@@ -1,12 +1,12 @@
 source 'https://rubygems.org'
 
-gem 'rails', '3.2.17'
+gem 'rails', '4.0.13'
+gem 'protected_attributes'
 gem 'color-tools', '>= 1.3.0'
 gem "will_paginate", "~> 3.0"
 gem "mysql2", "~> 0.3.11"
 gem 'dynamic_form'
 gem 'railroady'
-gem "therubyracer", "~> 0.11.4"
 gem 'jquery-rails'
 gem 'jquery-ui-rails'
 gem 'flot-rails'
@@ -16,14 +16,18 @@ gem 'ruby-units'
 gem 'gnuplot'
 gem 'axlsx_rails', "~> 0.3.0"
 
-# Gems used only for assets and not required
-# in production environments by default.
-group :assets do
-  gem 'sass-rails'
-  gem 'coffee-rails'
-  gem 'uglifier'
+gem 'sass-rails'
+gem 'coffee-rails'
+gem 'coffee-script-source', '1.8.0'
+gem 'uglifier'
+
+group :development do
+  gem 'quiet_assets'
+  gem 'pry-rails'
 end
 
 gem 'passenger'
 
-gem 'pry-rails', :group => :development
+gem 'thin'
+gem 'thin_service'
+gem 'actionpack-xml_parser'
