@@ -1,5 +1,4 @@
 include MangoModule
-include Rails.application.routes.url_helpers
 
 class EasyModel
   def self.production_note(params)
@@ -324,16 +323,16 @@ class EasyModel
         .where(recipes: {code: recipe_code})
         .order('orders.created_at')
         .group('recipes.version')
-        .pluck_all('recipes.version, MAX(orders.created_at) AS last_used_at')
+        .pluck('recipes.version, MAX(orders.created_at) AS last_used_at')
 
-      recipe_versions.sort! {|v1, v2| v1['last_used_at'] <=> v2['last_used_at']}
+      recipe_versions.sort! {|v1, v2| v1[1] <=> v2[1]}
 
       recipe_versions.each do |recipe_version|
         version_total = BatchHopperLot
           .joins({hopper_lot: {lot: {}},
                   batch: {order: {recipe: {}}}})
           .where(orders: {created_at: (start_date .. end_date + 1.day)},
-                 recipes: {code: recipe_code, version: recipe_version['version']})
+                 recipes: {code: recipe_code, version: recipe_version[0]})
           .sum(:amount)
 
         percentages = BatchHopperLot
@@ -342,7 +341,7 @@ class EasyModel
           .select("lots.ingredient_id,
                    SUM(batch_hoppers_lots.amount) / #{version_total} * 100 AS percentage")
           .where(orders: {created_at: (start_date .. end_date + 1.day)},
-                 recipes: {code: recipe_code, version: recipe_version['version']},
+                 recipes: {code: recipe_code, version: recipe_version[0]},
                  lots: {ingredient_id: ingredients_ids})
           .group('lots.ingredient_id')
           .reduce([{}, 0.0]) do |percentages, percentage|
@@ -353,8 +352,8 @@ class EasyModel
           end
         first_used_at = Order
           .joins(:recipe)
-          .where(['recipes.code = ? and recipes.version != ?', recipe_code, recipe_version['version']])
-          .where(['orders.created_at < ?', recipe_version['last_used_at']])
+          .where(['recipes.code = ? and recipes.version != ?', recipe_code, recipe_version[0]])
+          .where(['orders.created_at < ?', recipe_version[1]])
           .order('orders.created_at desc')
           .first.created_at.to_date
 
@@ -365,7 +364,7 @@ class EasyModel
         end
 
         row[:versions] << {
-          version: recipe_version['version'],
+          version: recipe_version[0],
           days: days,
           total: version_total / 1000,
           percentages: percentages[0],
@@ -422,9 +421,9 @@ class EasyModel
           .joins(:recipe)
           .where(created_at: week_range,
                  recipes: {code: recipe_code})
-          .pluck_all('DISTINCT recipes.version, recipes.id')
-          .sort_by { |hash| hash["version"] }
-          .map { |hash| {version: hash["version"], domain: domain, path: recipe_path(hash["id"]) } }
+          .pluck('DISTINCT recipes.version, recipes.id')
+          .sort_by { |hash| hash[0] }
+          .map { |hash| {version: hash[0], domain: domain, path: recipe_path(hash[1]) } }
       end
       row
     end
@@ -432,7 +431,7 @@ class EasyModel
   end
 
   def self.lot_transactions(start_date, end_date, lot_type, lot_code)
-    lot = lot_type == 1 ? Lot.find_by_code(lot_code) : ProductLot.find_by_code(lot_code)
+    lot = lot_type == 1 ? Lot.where(code: lot_code).first : ProductLot.where(lot: lot_code).first
     return nil if lot.nil?
 
     content = lot_type == 1 ? lot.ingredient : lot.product
@@ -526,7 +525,7 @@ class EasyModel
   end
 
   def self.hopper_transactions(hopper_id, start_datetime, end_datetime)
-    hopper = Hopper.find_by_id hopper_id, :include => :scale
+    hopper = Hopper.find_by id: hopper_id, :include => :scale
     return nil if hopper.nil?
 
     includes = {hopper_lot: {lot: {ingredient: {}}}, hopper_lot_transaction_type: {}, user: {}}
@@ -556,7 +555,8 @@ class EasyModel
   end
 
   def self.stats(start_date, end_date)
-    @orders = Order.find :all, :include=>{:order_stats => {}, :batch => {}, :recipe => {}}, :conditions=>['batches.start_date >= ? and batches.end_date <= ?', self.start_date_to_sql(start_date), self.end_date_to_sql(end_date)], :order=>['batches.start_date ASC']
+    @orders = Order.includes({:order_stats => {}, :batch => {}, :recipe => {}})
+      .where(['batches.start_date >= ? and batches.end_date <= ?', self.start_date_to_sql(start_date), self.end_date_to_sql(end_date)], :order=>['batches.start_date ASC'])
 
     data = self.initialize_data("Estadisticas de produccion")
     data['since'] = self.print_range_date(start_date)
@@ -616,7 +616,7 @@ class EasyModel
   end
 
   def self.order_stats(order_code)
-    order = Order.find_by_code order_code
+    order = Order.where(code: order_code).first
     return nil if order.nil?
 
     data = self.initialize_data('Estadisticas de orden')
@@ -771,7 +771,7 @@ class EasyModel
   end
 
   def self.alarms_per_order(order_code, alarm_type_id)
-    order = Order.find_by_code order_code
+    order = Order.where(code: order_code).first
     return nil if order.nil?
     alarms = []
     if alarm_type_id == 0
@@ -1039,7 +1039,7 @@ class EasyModel
   end
 
   def self.consumption_per_ingredient_per_orders(start_date, end_date, ingredient_id)
-    ingredient = Ingredient.find_by_id ingredient_id
+    ingredient = Ingredient.find_by id: ingredient_id
     return nil if ingredient.nil?
 
     batch_hopper_lots = BatchHopperLot
@@ -1157,7 +1157,8 @@ class EasyModel
   end
 
   def self.order_details(order_code)
-    @order = Order.find_by_code order_code, :include => {:batch => {:batch_hopper_lot => {:hopper_lot => {:hopper => {}, :lot=>{:ingredient=>{}}}}}, :recipe => {:ingredient_recipe => {:ingredient => {}}}, :medicament_recipe => {:ingredient_medicament_recipe => {:ingredient => {}}}, :product_lot => {:product => {}}, :client => {}}
+    @order = Order.includes({:batch => {:batch_hopper_lot => {:hopper_lot => {:hopper => {}, :lot=>{:ingredient=>{}}}}}, :recipe => {:ingredient_recipe => {:ingredient => {}}}, :medicament_recipe => {:ingredient_medicament_recipe => {:ingredient => {}}}, :product_lot => {:product => {}}, :client => {}})
+      .where(code: order_code).first
     return nil if @order.nil?
 
     ingredients = {}
@@ -1204,9 +1205,10 @@ class EasyModel
     ingredients.each do |key, value|
       unless details.has_key?(key)
         std_amount = value * n_batch
+        ingredient = Ingredient.where(code: key).first
         details[key] = {
-          'ingredient' => Ingredient.find_by_code(key).name,
-          'ingredient_id' => Ingredient.find_by_code(key).id,
+          'ingredient' => ingredient.name,
+          'ingredient_id' => ingredient.id,
           'lot' => "N/A",
           'hopper' => "N/A",
           'real_kg' => 0,
@@ -1268,10 +1270,11 @@ class EasyModel
   end
 
   def self.batch_details(order_code, batch_number)
-    order = Order.find_by_code(order_code)
+    order = Order.where(code: order_code).first
     return nil if order.nil?
 
-    batch = Batch.find :first, :include => {:order => {:recipe => {:ingredient_recipe => {:ingredient => {}}}, :medicament_recipe => {:ingredient_medicament_recipe => {:ingredient =>{}}}}}, :conditions => {:number => batch_number, :orders => {:code => order_code}}
+    batch = Batch.includes({:order => {:recipe => {:ingredient_recipe => {:ingredient => {}}}, :medicament_recipe => {:ingredient_medicament_recipe => {:ingredient =>{}}}}})
+      .where({:number => batch_number, :orders => {:code => order_code}}).first
     return nil if batch.nil?
 
     data = self.initialize_data('Detalle de Batch')
@@ -1282,7 +1285,9 @@ class EasyModel
     data['end_date'] = batch.end_date.strftime("%d/%m/%Y %H:%M:%S")
     data['results'] = []
 
-    batch_hopper_lots = BatchHopperLot.find :all, :include => {:hopper_lot => {:hopper => {}, :lot => {:ingredient => {}}}}, :conditions => {:batch_id => batch.id}, :order=>['ingredients.code ASC']
+    batch_hopper_lots = BatchHopperLot.includes({:hopper_lot => {:hopper => {}, :lot => {:ingredient => {}}}})
+      .where({:batch_id => batch.id})
+      .order('ingredients.code ASC')
 
     ingredients = {}
     batch.order.recipe.ingredient_recipe.each do |ir|
@@ -1329,7 +1334,7 @@ class EasyModel
     ingredient_inclusion = params[:ingredient_inclusion] == '1'
     by_lots = params[:by_lots_recipe] == '1'
 
-    recipe = Recipe.find_by_code recipe_code
+    recipe = Recipe.where(code: recipe_code).first
     return nil if recipe.nil?
 
     data = self.initialize_data('Consumo por Receta')
@@ -1489,7 +1494,7 @@ class EasyModel
   end
 
   def self.stock_adjustments(start_date, end_date)
-    transaction_types = TransactionType.find :all
+    transaction_types = TransactionType.all
     adjustment_type_ids = []
     transaction_types.each do |ttype|
       unless ttype.code.match(/(?i)AJU/).nil?
@@ -1497,10 +1502,10 @@ class EasyModel
         adjustment_type_ids << ttype.id
       end
     end
-    return nil if adjustment_type_ids.length.zero?
+    return nil if adjustment_type_ids.empty?
 
-    adjustments = Transaction.find :all, :conditions => {:transaction_type_id => adjustment_type_ids, :created_at => (start_date)..((end_date) + 1.day)}, :order=>['created_at DESC']
-    return nil if adjustments.length.zero?
+    adjustments = Transaction.where({:transaction_type_id => adjustment_type_ids, :created_at => (start_date)..((end_date) + 1.day)}, :order=>['created_at DESC'])
+    return nil if adjustments.empty?
 
     data = self.initialize_data('Ajustes de Existencias')
     data['since'] = self.print_range_date(start_date)
@@ -1543,12 +1548,12 @@ class EasyModel
   end
 
   def self.lots_incomes(start_date, end_date)
-    income_type = TransactionType.find :first, :conditions => {:code => 'EN-COM'}
+    income_type = TransactionType.where({:code => 'EN-COM'}).first
     "Income code found: " + income_type.code
     return nil if income_type.nil?
 
-    incomes = Transaction.find :all, :conditions => {:transaction_type_id => income_type, :created_at => (start_date)..((end_date) + 1.day)}, :order=>['created_at DESC']
-    return nil if incomes.length.zero?
+    incomes = Transaction.find.where({:transaction_type_id => income_type, :created_at => (start_date)..((end_date) + 1.day)}, :order=>['created_at DESC'])
+    return nil if incomes.empty?
 
     data = self.initialize_data('Entradas de Materia Prima')
     data['since'] = self.print_range_date(start_date)
@@ -1607,10 +1612,10 @@ class EasyModel
       lots = lots.order('products.code, products_lots.code asc')
     end
     lots.each do |lot|
-      transaction = Transaction.first :conditions => [
+      transaction = Transaction.where([
         'notified = true and created_at < ? and content_type = ? and content_id = ? ',
         date.strftime("%Y-%m-%d %H:%M:%S"), content_type, lot.id
-      ], :order => ['created_at desc']
+      ]).order('created_at desc').first
       if transaction
         data['results'] << {
           'code' => lot.code,
@@ -1652,10 +1657,10 @@ class EasyModel
 
     lots.each do |l|
       key = l.get_content.code
-      transaction = Transaction.first :conditions => [
+      transaction = Transaction.where([
         'notified = true and created_at < ? and content_type = ? and content_id = ? ',
         date.strftime("%Y-%m-%d %H:%M:%S"), content_type, l.id
-      ], :order => ['created_at desc']
+      ]).order('created_at desc').first
       if transaction
         if results.has_key?(key)
           results[key]['stock'] += transaction.stock_after
@@ -1731,8 +1736,8 @@ class EasyModel
       conditions = {:transaction_type_id=>5, :transactions=>{:document_number=>doc_number}, :created_at => (start_date)..((end_date) + 1.day)}
     end
 
-    dispatches = Transaction.find :all, :conditions => conditions, :order=>['created_at DESC']
-    return nil if dispatches.length.zero?
+    dispatches = Transaction.where(conditions).order('created_at DESC')
+    return nil if dispatches.empty?
 
     data = self.initialize_data('Despacho de producto terminado')
     data['since'] = self.print_range_date(start_date)
@@ -1761,7 +1766,7 @@ class EasyModel
   end
 
   def self.production_per_recipe(start_date, end_date, recipe_code)
-    recipe = Recipe.find_by_code recipe_code
+    recipe = Recipe.where(code: recipe_code).first
     return nil if recipe.nil?
 
     data = self.initialize_data('Produccion por Receta')
