@@ -1,11 +1,13 @@
 class Carrier < ActiveRecord::Base
+  attr_protected :id
+
   has_many :trucks
   accepts_nested_attributes_for :trucks
 
   validates_presence_of :name
   validates_length_of :name, :within => 3..40
   before_save :create_code
-  
+
   def create_code
     unless self.id
       last = Carrier.last
