@@ -1,4 +1,6 @@
 class HopperLotTransaction < ActiveRecord::Base
+  attr_protected :id
+
   belongs_to :hopper_lot_transaction_type
   belongs_to :hopper_lot
   belongs_to :user
@@ -30,8 +32,7 @@ class HopperLotTransaction < ActiveRecord::Base
   end
 
   def get_sign
-    hltt = HopperLotTransactionType.find(self.hopper_lot_transaction_type_id)
-    hltt.sign
+    HopperLotTransactionType.find(self.hopper_lot_transaction_type_id).sign
   end
 
   def do_stock_update
