@@ -1,4 +1,6 @@
 class IngredientRecipe < ActiveRecord::Base
+  attr_protected :id
+
   belongs_to :ingredient
   belongs_to :recipe
 
@@ -17,12 +19,12 @@ class IngredientRecipe < ActiveRecord::Base
   end
 
   def validate_existence
-    if IngredientRecipe.find(:first, :conditions => ["ingredient_id = ? and recipe_id = ?", self.ingredient_id, self.recipe_id])
+    if IngredientRecipe.where(["ingredient_id = ? and recipe_id = ?", self.ingredient_id, self.recipe_id]).first
       errors.add_to_base("ingredient already exist")
     end
     return (errors.size > 0) ? false : true
   end
-  
+
   def get_percentage
     amount * 100 / self.recipe.get_total
   end
