@@ -1,7 +1,9 @@
 class MedicamentRecipe < ActiveRecord::Base
+  attr_protected :id
+
   has_many :ingredient_medicament_recipe, :dependent => :destroy
   has_many :order
-  
+
   validates_uniqueness_of :code
   validates_presence_of :name, :code
   validates_length_of :name, :within => 3..40
@@ -11,11 +13,6 @@ class MedicamentRecipe < ActiveRecord::Base
   end
 
   def get_total
-    total = 0
-    ingredients = IngredientMedicamentRecipe.where(:medicament_recipe_id => self.id)
-    ingredients.each do |i|
-      total+= i.amount
-    end
-    return total
+    ingredient_medicament_recipe.pluck(:amount).reduce { |total, amount| total + amount }
   end
 end
