@@ -1,4 +1,6 @@
 class Permission < ActiveRecord::Base
+  attr_protected :id
+
   has_many :permission_role
 
   after_save :associate_role
@@ -8,7 +10,7 @@ class Permission < ActiveRecord::Base
 
   MODULES = ['batches', 'orders', 'recipes', 'hoppers', 'batches_hopper_lot', 'transactions', 'lots',
     'product_lots', 'ingredients', 'ingredients_recipes', 'products', 'clients', 'factories', 'transaction_types',
-    'schedules', 'users', 'roles', 'permissions', 'reports', 'configuration', 'medicament_recipes', 'drivers', 'carriers', 
+    'schedules', 'users', 'roles', 'permissions', 'reports', 'configuration', 'medicament_recipes', 'drivers', 'carriers',
     'trucks', 'tickets', 'alarm_types', 'parameters', 'parameter_types', 'parameter_lists', 'scales',
     'lot_parameters', 'lot_parameter_types', 'lot_parameter_lists', 'settings']
   # Permission actions
