@@ -1,9 +1,11 @@
 class PermissionRole < ActiveRecord::Base
+  attr_protected :id
+
   belongs_to :role
   belongs_to :permission
 
   def self.find_with_permissions(role_id, controller, mode)
-    return find :all, :conditions => {:role_id => role_id, :permissions => {:module => controller, :mode=>mode}}, :include => [:permission]
+    return includes(:permission).where({:role_id => role_id, :permissions => {:module => controller, :mode=>mode}})
   end
 
   def self.generate_god_role
@@ -21,8 +23,8 @@ class PermissionRole < ActiveRecord::Base
   def self.generate_scada_role
     scada_role = Role.find(2)
     scada_role.permission_role.clear
-    permissions = Permission.find :all, :conditions=>["module = 'batches'"]
-    permissions << Permission.find(:first, :conditions=>["module = 'transactions' AND action = 'SA-CSM'"])
+    permissions = Permission.where(["module = 'batches'"])
+    permissions << Permission.where(["module = 'transactions' AND action = 'SA-CSM'"]).first
     permissions.each do |perm|
       perm_role = PermissionRole.new
       perm_role.permission_id = perm.id
