@@ -1,3 +1,4 @@
 class PreselectedIngredientId < ActiveRecord::Base
-  # attr_accessible :title, :body
+  attr_protected :id
+
 end
