@@ -1,11 +1,17 @@
 class Scale < ActiveRecord::Base
+  attr_protected :id
+
   has_many :hoppers
   validates :name, presence: true
   validates :name, uniqueness: true
-  validates :not_weighed, uniqueness: { if: :not_weighed }
+  validate :not_weighed_uniqueness
   validates :maximum_weight, :minimum_weight, presence: { unless: :not_weighed }
   validates :minimum_weight, numericality: { greater_than_or_equal_to: 0, less_than: :maximum_weight, allow_nil: true }
   validates :maximum_weight, numericality: { greater_than: 0, allow_nil: true }
+
+  def not_weighed_uniqueness
+    errors.add(:not_weighed, :taken) if not_weighed && Scale.where(not_weighed: true).any?
+  end
 
   def self.get_all
     scales = Scale.order('not_weighed')
@@ -19,4 +25,25 @@ class Scale < ActiveRecord::Base
       end
     return scales, hoppers_below_minimum
   end
+
+  def self.get_hoppers_ingredients
+    scales = Scale.where(not_weighed: false)
+      .pluck(:id, :name, :minimum_weight, :maximum_weight)
+      .map do |scale|
+        {id: scale[0], name: scale[1], minimum_weight: scale[2], maximum_weight: scale[3]}
+      end
+    scales.each do |scale|
+      scale[:hoppers] = HopperLot.includes({lot: {ingredient: {}}, hopper: {scale: {}}})
+        .where(active: true)
+        .where(hoppers: {scale_id: scale[:id]})
+        .pluck('hoppers.number', 'hoppers.name AS hopper_name', 'ingredients.name AS ingredient_name', 'ingredients.code AS ingredient_code')
+        .map do |hopper|
+          {number: hopper[0], name: hopper[1], ingredient_name: hopper[2], ingredient_code: hopper[3]}
+        end
+    end
+
+    scales
+  end
 end
+
+
