@@ -1,4 +1,6 @@
 class TransactionType < ActiveRecord::Base
+  attr_protected :id
+
   has_many :transactions
 
   validates_uniqueness_of :code
@@ -15,7 +17,7 @@ class TransactionType < ActiveRecord::Base
 
   def destroy_permission
     begin
-      p = Permission.find :first, :conditions=>{:module=>'transactions', :action=>self.code, :mode=>'module'}
+      p = Permission.where({:module=>'transactions', :action=>self.code, :mode=>'module'}).first
       p.destroy
     rescue
     end
