@@ -18,8 +18,9 @@ Mango::Application.routes.draw do
   match 'recipes/:id/deactivate' => "recipes#deactivate", :via => :get, :as => 'deactivate_recipe'
   match 'roles/:id/clone' => "roles#clone", :via => :get, :as => 'clone_role'
   match 'tickets/:id/print' => "tickets#print", :via => :get, :as => 'print_ticket'
-  match 'tickets/:id/notify' => "tickets#notify", :via => :get, :as => 'notify_ticket'
-  match 'orders/:id/repair' => "orders#repair", :via => :get, :as => 'repair_order'
+  match 'tickets/:id/notify' => "tickets#notify", :via => :get#, :as => 'notify_ticket'
+  match 'tickets/:id/notify' => "tickets#do_notify", :via => :post, :as => "notify_ticket"
+  match 'orders/:id/repair' => "orders#repair", :via => :get#, :as => 'repair_order'
   match 'orders/:id/repair' => "orders#do_repair", :via => :put, :as => 'repair_order'
   match 'orders/:id/print' => "orders#print", :via => :get, :as => 'print_order'
   match 'orders/:id/generate_transactions' => "orders#generate_transactions", :via => :put, :as => 'generate_transactions_order'
@@ -29,13 +30,15 @@ Mango::Application.routes.draw do
   match 'orders/close' => "orders#close", :via => :post, :as => 'order_close'
   match 'orders/create_order_stat' => "orders#create_order_stat", :via => :post, :as => 'create_order_stat'
   match 'orders/update_order_area' => "orders#update_order_area", :via => :post, :as => 'update_order_area'
-  match 'orders/:id/notify' => "orders#notify", :via => :get, :as => 'notify_order'
+  match 'orders/:id/notify' => "orders#notify", :via => :get#, :as => 'notify_order'
   match 'orders/:id/notify' => "orders#do_notify", :via => :post, :as => 'notify_order'
+  match 'orders/open' => "orders#open", :via => :get, :as => 'open_orders'
+  match 'orders/validate' => "orders#validate", :via => :get, :as => 'validate_order'
   match 'sessions/not_implemented' => "sessions#not_implemented", :via => :get, :as => "not_implemented"
-  match 'lots/:id/adjust' => "lots#adjust", :via => :get, :as => 'adjust_lot'
+  match 'lots/:id/adjust' => "lots#adjust", :via => :get#, :as => 'adjust_lot'
   match 'lots/:id/adjust' => "lots#do_adjust", :via => :put, :as => 'adjust_lot'
   match 'lots/get_all' => "lots#get_all", :via => :get
-  match 'product_lots/:id/adjust' => "product_lots#adjust", :via => :get, :as => 'adjust_product_lot'
+  match 'product_lots/:id/adjust' => "product_lots#adjust", :via => :get#, :as => 'adjust_product_lot'
   match 'product_lots/:id/adjust' => "product_lots#do_adjust", :via => :put, :as => 'adjust_product_lot'
   match 'product_lots/get_all' => "product_lots#get_all", :via => :get
   match 'product_lots/by_recipe' => "product_lots#by_recipe", :via => :get
@@ -45,6 +48,7 @@ Mango::Application.routes.draw do
   match 'scales/:scale_id/hoppers/:id/change' => "hoppers#do_change", :via => :post
   match 'scales/:scale_id/hoppers/:id/adjust' => "hoppers#do_adjust", :via => :post
   match 'scales/:scale_id/hoppers/:id/change_factory_lots' => "hoppers#do_change_factory_lots", :via => :put
+  match 'scales/hoppers_ingredients' => "scales#hoppers_ingredients", :via => :get
   match 'settings.json' => "settings#show", via: :get
   match 'settings/features.json' => "settings#features", via: :get
   match 'settings' => "settings#edit", :via => :get, :as => "settings"
@@ -54,8 +58,7 @@ Mango::Application.routes.draw do
   match 'batches/:batch_id/batches_hopper_lot/:id' => "batches_hopper_lot#destroy", :via => :delete, :as => "batch_hopper_lot"
   match 'tickets/:id/repair' => "tickets#repair", :via => :get, :as => "repair_ticket"
   match 'tickets/:id/do_repair' => "tickets#do_repair", :via => :post, :as => "do_repair_ticket"
-  match 'tickets/:id/notify' => "tickets#notify", :via => :get, :as => "notify"
-  match 'tickets/:id/notify' => "tickets#do_notify", :via => :post, :as => "notify_ticket"
+
   # Reports
   match 'reports' => "reports#index", :via => :get, :as => "reports"
   match 'reports/recipes' => "reports#recipes", :via => :post, :as => "recipes_report"
@@ -98,6 +101,7 @@ Mango::Application.routes.draw do
     :transaction_types, :product_lots, :permissions, :drivers, :carriers, :trucks, :alarm_types, :parameter_types, :order_stat_types,
     :lot_parameter_types, :product_lot_parameter_types, :ingredient_parameter_type_ranges, :product_parameter_type_ranges
   resources :transactions, only: [:index, :new, :create]
+  resources :alarms, only: [:create]
   resources :document_types, only: [:index]
   resources :tickets, except: :new
 
@@ -138,5 +142,5 @@ Mango::Application.routes.draw do
 
   root :to => "sessions#index"
 
-  match "/sessions/show" => "sessions#show", as: 'dashboard'
+  match "/sessions/show" => "sessions#show", via: :get, as: 'dashboard'
 end
